Move summary and tradeinfo to tradeinfopanel component
