+++ src/botPage/view/View.js
@@ -23,8 +23,7 @@ import logHandler from './logger';
 import Tour from './tour';
 import OfficialVersionWarning from './react-components/OfficialVersionWarning';
 import updateLogTable from './updateLogTable';
-import updateTradeTable, { open as rerenderTradeTable } from './updateTradeTable';
-import updateSummary from './updateSummary';
+import TradeInfoPanel from './TradeInfoPanel';
 
 let realityCheckTimeout;
 
@@ -167,6 +166,7 @@ export default class View {
                         />,
                         $('#footer')[0]
                     );
+                    ReactDOM.render(<TradeInfoPanel />, $('#summaryPanel')[0]);
                     resolve();
                 });
             });
@@ -302,7 +302,6 @@ export default class View {
         });
         const showSummary = () => {
             $('#summaryPanel').dialog('open');
-            rerenderTradeTable();
         };
         $('#showSummary').click(showSummary);
 
@@ -417,7 +416,6 @@ export default class View {
         });
 
         globalObserver.register('bot.info', info => {
-            updateSummary(info);
             if ('profit' in info) {
                 const token = $('.account-id').first().attr('value');
                 const user = getToken(token);
@@ -426,12 +424,6 @@ export default class View {
                     profit  : info.profit,
                     contract: info.contract,
                 });
-            }
-        });
-
-        globalObserver.register('bot.contract', c => {
-            if (c) {
-                updateTradeTable(c);
             }
         });
     }
